Move theme to top level
Why have a utils folder if this is the only thing in it?
